Bump to 2.1; I think it is ready

+++ release.sh
@@ -1,6 +1,12 @@
 #!/bin/sh -e
 set -e
 
+
+if [ ! -f common.h ];then
+	echo 'wrong pwd?'
+	exit 1;
+fi
+
 VERSION=$(fgrep VERSION common.h |cut -f3 |cut -d' ' -f2 |tr -d '"')
 
 rm -Rf ../kc-"${VERSION}"
